+++ Common/src/main/java/me/frankv/jmi/mixin/UpdateFromServerTaskMixin.java
@@ -3,7 +3,7 @@ package me.frankv.jmi.mixin;
 import dev.ftb.mods.ftbchunks.client.map.MapDimension;
 import dev.ftb.mods.ftbchunks.client.map.UpdateChunkFromServerTask;
 import dev.ftb.mods.ftbchunks.net.SendChunkPacket;
-import me.frankv.jmi.jmoverlay.ftbchunks.ClaimedChunkPolygon;
+import me.frankv.jmi.compat.ftbchunks.claimedchunksoverlay.ClaimedChunksOverlay;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -16,14 +16,20 @@ import java.util.UUID;
 @Mixin(UpdateChunkFromServerTask.class)
 public class UpdateFromServerTaskMixin {
 
-    @Shadow(remap = false) @Final private MapDimension dimension;
+    @Shadow(remap = false)
+    @Final
+    private MapDimension dimension;
 
-    @Shadow(remap = false) @Final private SendChunkPacket.SingleChunk chunk;
+    @Shadow(remap = false)
+    @Final
+    private SendChunkPacket.SingleChunk chunk;
 
-    @Shadow(remap = false) @Final private UUID teamId;
+    @Shadow(remap = false)
+    @Final
+    private UUID teamId;
 
     @Inject(method = "runMapTask", at = @At("RETURN"), remap = false)
     private void onRunMapTask(CallbackInfo ci) {
-        ClaimedChunkPolygon.INSTANCE.addToQueue(dimension, chunk, teamId);
+        ClaimedChunksOverlay.INSTANCE.addToQueue(dimension, chunk, teamId);
     }
 }